// src/interfaces/interfaces.ts
export interface IMainAcc {
    status?: string;
    readonly proxy: string;
    readonly country: string;
    readonly login: string;
    readonly pass: string;
    readonly email: string;
    readonly FName: string;
    readonly LName: string;
    readonly useragent: string;
    readonly gender: string;
    readonly birthday: string;
    readonly payeer: string;
    readonly UserInfo: string;
}